Fixed potential incorrect remembering bug
Given rules:
T = A | B | C | D
S = C | D

If, at some position, S is checked and C is parsed, then the parser
would remember C. Supposing the containing rule failed, it might test T
at the same position. The old behaviour was simply to return C, but it
is possible that A could be correctly parsed. Moreover, if S failed, T
will reattempt parsing of C and D at this position.

The included changes memorises rules, not results. Therefore, when T is
parsed, it will attempt A and B first. If it gets to C, it will check
the memory and return the result of C.

This is slightly more costly, memory-wise, but much more correct and
avoids potential reparsing.